Fixed saga code. Removed unnecessary babel dependencies and extra files.
