+++ package.json
@@ -29,11 +29,8 @@
     "@babel/core": "^7.0.0",
     "@babel/plugin-proposal-class-properties": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
-    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
     "docsify-cli": "^4.2.1",
     "eslint": "^5.6.1",
     "eslint-config-standard": "^12.0.0",
@@ -43,7 +40,7 @@
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-standard": "^4.0.0",
     "jest": "^23.6.0",
-    "parcel-bundler": "^1.10.1"
+    "parcel-bundler": "^1.10.3"
   },
   "scripts": {
     "test": "jest",
