--- problems/10/chat/olaf/.babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": [
-    ["env", {
-      "targets": {
-        "browsers": "last 1 chrome versions"
-      }
-    }]
-  ],
-  "plugins": ["transform-class-properties"]
-}
