--- problems/10/chat/olaf/.editorconfig
@@ -1,15 +1,0 @@
-root = true
-
-[*]
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-charset = utf-8
-
-[{*.js,*.css}]
-indent_style = space
-indent_size = 2
-
-[{package.json,*.yml,*.cjson}]
-indent_style = space
-indent_size = 2
