--- problems/10/chat/olaf/.eslintrc.js
@@ -1,10 +1,0 @@
-module.exports = {
-  "parser": "babel-eslint",
-  "extends": "standard",
-  "plugins": [
-    "babel"
-  ],
-  "rules": {
-    "strict": 0
-  }
-};
