--- problems/10/chat/olaf/README.md
@@ -1,9 +1,0 @@
-# olaf
-A P2P chat
-
-## Commands
-Command                | Description                                      |
------------------------|--------------------------------------------------|
-`$ npm start`          | Start the development server
-`$ npm test`           | Lint, validate deps & run tests
-`$ npm run build`      | Compile all files into `dist/`
