--- problems/10/chat/olaf/manifest.json
@@ -1,14 +1,0 @@
-{
-  "name": "olaf",
-  "short_name": "olaf",
-  "description": "A P2P chat",
-  "start_url": "/",
-  "display": "standalone",
-  "background_color": "#000",
-  "theme_color": "#000",
-  "icons": [{
-    "src": "/assets/icon.png",
-    "type": "image/png",
-    "sizes": "512x512"
-  }]
-}
